Move form filling logic to form_filling.py
- Add typing where applicable
